scraper/requirements.txt
selenium
bs4
google-cloud-datastore
